Refactor tauri.js to move command code outside of bin files (#140)
* refactor(tauri.js): move command code into helpers

* reafactor(tauri.js): move command helpers to api dir

* fix(tauri.js/tests): import tauricon from proper dir

* fix(tauri.js/tests): add api dir to moduleNameMapper
