+++ cli/tauri.js/api/init.js
@@ -1,0 +1,9 @@
+const { inject } = require('../template')
+
+module.exports = args => {
+  return inject(args.directory, 'all', {
+    force: args.force,
+    logging: args.logging,
+    tauriPath: args.tauriPath
+  })
+}
