+++ cli/tauri.js/bin/tauri-build.js
@@ -1,5 +1,4 @@
-const
-  parseArgs = require('minimist')
+const parseArgs = require('minimist')
 
 const argv = parseArgs(process.argv.slice(2), {
   alias: {
@@ -21,17 +20,6 @@ if (argv.help) {
   process.exit(0)
 }
 
-const { tauriDir } = require('../helpers/app-paths')
-const Runner = require('../runner')
-const tauri = new Runner({ modeDir: tauriDir })
-const tauriConfig = require('../helpers/tauri-config')({
-  ctx: {
-    debug: argv.debug,
-    prod: true
-  }
-})
+const build = require('../api/build')
 
-require('../generator').generate(tauriConfig.tauri)
-require('../entry').generate(tauriDir, tauriConfig)
-
-tauri.build(tauriConfig)
+build({ ctx: { debug: argv.debug } })
